Run Kotlin test on Postgres
Postgres started using TestContainers
