+++ server-kotlin/build.gradle
@@ -46,6 +46,7 @@ dependencies {
     testCompile "io.ktor:ktor-server-tests:$ktor_version"
     testCompile "com.h2database:h2:1.4.200"
     testCompile "net.javacrumbs.json-unit:json-unit-assertj:2.14.0"
+    testCompile "org.testcontainers:postgresql:1.13.0"
 
     // google docs API
     compile 'com.google.api-client:google-api-client:1.30.1'
